cleanup of ntske package
